refactor(mir-lower): generalize target-stable aggregate storage
